+++ back/package.json
@@ -3,7 +3,6 @@
   "version": "0.0.1",
   "description": "",
   "author": "",
-  "type": "module",
   "private": true,
   "license": "UNLICENSED",
   "scripts": {
